Remove unnecessary main view
