+++ lib/assets/javascripts/cartodb/new_keys/entry.js
@@ -1,14 +1,15 @@
 var $ = require('jquery');
 var cdb = require('cartodb.js');
 var trackJsErrors = require('new_common/track_js_errors');
-var MainView = require('new_keys/main_view');
+// var HeaderView = require('nnew_common/header_viewew');
+var SupportView = require('new_common/support_view');
 
 if (window.trackJs) {
   trackJsErrors(window.trackJs, window.user_data.username);
 }
 
 /**
- * Entry point for the new dashboard, bootstraps all dependency models and application.
+ * Entry point for the new keys, bootstraps all dependency models and application.
  */
 $(function() {
   cdb.init(function() {
@@ -19,15 +20,41 @@ $(function() {
       cdb.config.set('url_prefix', cdb.config.organizationUrl());
     }
 
-    var currentUser = new cdb.admin.User(user_data);
+    var user = new cdb.admin.User(user_data);
 
-    var keys = new MainView({
-      el: document.body,
-      user: currentUser,
-      config: config
+    // var headerView = new HeaderView({
+    //   el:         this.$('#header'), //pre-rendered in DOM by Rails app
+    //   model:      this.user,
+    //   router:     this.router,
+    //   localStorage: this.localStorage
+    // });
+    // headerView.render();
+
+    var supportView = new SupportView({
+      el: $('#support-banner'),
+      user: user
     });
-    
-    window.keys = keys;
+    supportView.render();
+
+    // TODO: replace .Form-copy => .js-copy-api-key
+    $('.Form-copy').each(function() {
+      // Copy
+      $(this).zclip({
+        path: cdb.config.get('assets_url') + "/flash/ZeroClipboard.swf",
+        copy: function(){
+          // TODO: replace with .js-input
+          return $(this).parent().find(".Form-input").val();
+        }
+      });
+
+      // Tooltip
+      var tooltip = new cdb.common.TipsyTooltip({
+        el: $(this),
+        title: function() {
+          return $(this).data('title');
+        }
+      });
+    });
   });
 
 });
